refactor : 타입 수정
