+++ apis/auth/type.ts
@@ -1,0 +1,6 @@
+export type LoginResponse = {
+  errorCode: string;
+  userId: number;
+  authority: string;
+  token: string;
+};
